Attempting to fix M-/...
Old emacs would auto-complete to case of the completion candidate,
ignoring the case of what I had already typed... I want that back.

Right now it does the opposite, so this is the first attempt to fix.

Can't currently tell if it works or not... May have a stale
auto-complete cache or something.