mis bugfixes
Fixed `mis/comment/header' to work with indentation.

Added `existing' option to `mis/string/indent' for "don't actually
indent anything but use existing indentation for calculating stuff".
  - Also added to the rest of mis/args/string.el funcs that deal with
    indentation.

Fixed `mis/comment/wrap' to know about/respect indentation. And to not
trim the indentation if the string has been asked to be trimmed.

Fixed a lot of test calls to actually use the params correctly.
  - lists of mlists are not completely obvious to use, apparently.