Fixed org-todo-keyword-faces, removed org-roam.
Fixed bug with `cons' vs `list' in `org-todo-keyword-faces' by changing
all of them to be lists now.

Updated setting `spy/config.root' by tweaking the code that got what dir
the root is. Now it can function properly if it's in a subdirectory of
the doom user directory.

Not using org-roam correctly/at-all currently, so disabling.