Change output default value setup.
Changed how the output vars' default values work. Putting to a symbol
property and getting from that property doesn't work reliably in some
tests?

Changed to have defconsts of the default values, initialize the variable
from (a copy of) the const, and reset the variable to (a copy of) the
const again.