Refactor imp defcustoms.
Changed most `defcustom' imp variables to be defconst instead.

imp is intended to be used by libs, packages, and user configs alike so
changing how it parses strings to understand the structuring of features
is not something it should be supporting.

That made sense when I was gonna use it just for my config, but it
doesn't make sense now that I'm starting to think of my config as a
bunch of custom "packages" plus... my config.