Taskspace: Fix bug introduced in recent bugfix
Fix bug in "config/taskspace.el". Checks were added in order to avoid
loading taskspace if the `jerky` variables were not set for it. However,
the check included variables that would be set inside the code it
guards.

Removed the general check. Added checks for each domain (`:home` and
`:work`) that do not over-reach. This allows either/both domains to be
added to `taskspace` depending on what exists in `jerky`.

Also allows Taskspace to always be minimally set up if nothing is
defined (always creates the default domain).

Removed feature flag check in "taskspace/init.el". No point in it; why
would you want the taskspace module but not any actual taskspace code
loaded?