Fix backtraces when erroring during start up.
`innit:load:ordered:files` was using `condition-case` to catch errors
for messaging about what file it was loading for additional context.
This causes the backtrace to lose the actual stack and create a new one
in the `condition-case` handler where it was "rethrown".

Fix to use `unwind-protect` instead of `condition-case` so we can check
if normal or interrupted exit and log out or info without affecting the
error signal's stack.