Update parsing to support parsed Mis ASTs.
Update parsing to not mis-parse a Mis AST.

For example:
  (mis:comment (mis:line "-"))

The line will get parsed first, and an AST returned, and then the
comment will get parsed. We need the comment to not misunderstand the
line's AST when parsing.

Update `int<mis>:parse` to just pass ASTs through unchanged, as long as
they're an expected category.
  - Change internal categories to have `:mis' prefix (`:mis:format') so
    they are different from the input categories.

Update `int<mis>:valid:syntax?` to have a `no-error?` mode for support
of parsing changes.