Split `elisp:cmd/with-args` into two macros.
Split into:
  - `elisp:cmd/prefix` - for only a prefix arg.
  - `elisp:cmd/args`   - for args, no prefix arg.

Split due to Doom's `cmd!!` (which was `elisp:cmd/with-args`) not being
intuitive. I was messing up how to call it even after I knew what was
wrong and how to fix it.

Now the keybinds that want a prefix use one and the keybinds that want
args use the other and there is nothing that wants both.