Tweak marginalia widths
Bump `marginalia-field-width` up to 310.

Use advice to fix `marginalia--affixate` to use frame width instead of
window width for its max width calculations. The minibuffer is the
width of the entire frame, so I don't know what they were thinking
there...

Use advice to fix `marginalia-annotate-buffer` so that it gives a more
usable amount of width to the buffer file path field.