Add super basic template module to core
Add `str:propertize:remove` function for deleting text properties from
a string.

Add `core/modules/emacs/template` to the core modules. It is a
non-interactive template/snippet library, as opposed to the
interactive YASnippet.

Current it is super basic - only good enough to be maybe sort of
useful in special cases, to start off.