Fix some `all-the-icons` bugs
Bug 1
-----

Remove, for now, the two part `defun` + override `define-advice`. It's
for being able to use those functions still, even when `all-the-icons`
shouldn't load because non-gui mode.

However...

It put past my understanding something important: getting
`all-the-icons` and those finalized versions actually loaded when
they're wanted.

So for now just `defun` the helpers in the `use-package` `:config`
section, and tell `use-package` about them via an `:autoload` section.

Bug 2
-----

Give the helper functions two different faces: one for the text and
one for the icon. This should probably have been in the (a?) previous
commit?