Start of feature that was too much for now
Made some functions for gathering up all templates for a buffer. The
idea was to use it to create a hydra or something of what all a buffer
can do. However a template isn't complete without its parameters, so
we can't really do what I was trying right now.

Abandon for now. Think about how to do correctly?