Fix bug in `path:uniquify`
New `str:normalize` funcs don't filter out buffer params, I guess? But
the actual bugs is that a function was getting its params in the wrong
order and another function was being called with a param it
shouldn't've ever been.

Together they used to canceled out, but recent changes upset that.

Fix to not be bugged.