Add `nerd-icons` package
Rename "all-the-icons.el" to "fonts-and-icons.el", as it's a good
place to put `nerd-icons` config.

Fix `all-the-icons` config to autoinstall properly. That is, only
autoinstall when its fonts are not installed. It had regressed some
time in the past few months to installing every start-up.

Add `nerd-icons` config to "fonts-and-icons.el". Use the autoinstall
fix that `all-the-icons` just got and do `nerd-icons` autoinstall the
same.

Well, almost the same. Font names are more difficult for `nerd-icons`
than `all-the-icons`, so right now I only know what this Ubuntu has as
its nerd font.