Fix "Error reading undo-tree history"
Not exactly a fix; more of an "avoid".

Remap some keybinds so as to try to not use normal exit buffer
functions for quitting the `undo-tree` visualizer; try to use the
special functions (e.g. `undo-tree-visualizer-quit`) instead.

---

Also switch some other consts over to using the `unit:byte` helper
instead of manually mathing.