Fix snippet expansion bug
Bug:
----

If there is no (installed) mode matching the mode of an org-mode
`src` snippet, an error is raised:

> org-edit-src-code: No such language mode: rg-mode

This causes the snippet to be aborted before anything useful happens,
and since I have fancy history on that snippet, the next snippet(s)
will also be aborted until the history is manually cleared.

Fix:
----

New function `mantel:yas:indent-according-to-mode/safe` wraps
`indent-according-to-mode` and converts the error signal to a message.

New advice on `yas--indent-region` lexically binds that safe indent
function to `indent-according-to-mode` so that it doesn't error and
the snippet can continue on unimpeded.