Fix impatient-mode previews
Switch to an `impatient-mode` managed by `straight`, so we can use a
fork that we've just done a bugfix for.

Update `mantle:cmd:markdown:preview-live/start` to use
`imp-visit-buffer` to go directly to buffer preview instead of going
to list of buffers and making user click the one they wanted all
along.